# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud Backup Server API
  description: |
    REST API for the Cloud Backup Server.

    ## Authentication

    This API uses a **two-token** scheme with **HttpOnly cookie delivery** as the primary
    transport (preferred for browser and Electron clients):

    - **Access token** — short-lived JWT (1 minute). Set as the `access_token` HttpOnly cookie
      (`SameSite=Lax`, `Path=/`). Also accepted as `Authorization: Bearer <token>` for
      non-browser clients (e.g. Electron) that cannot use cookies.
    - **Refresh token** — long-lived opaque token (30 days). Set as the `refresh_token` HttpOnly
      cookie (`SameSite=Lax`, `Path=/api/auth/refresh`). Can also be sent in the JSON request body
      for clients that cannot use cookies. Only the SHA-256 hash is stored in the database; the
      raw token is never persisted.

    Tokens are also returned in the JSON response body on login/register/refresh for clients
    that cannot use cookies (e.g. Electron from a `file://` origin). Insomnia users should
    rely on the cookie jar instead — run login once and all requests are authenticated automatically.

    ### Token rotation
    Every call to `/api/auth/refresh` issues a *new* refresh token and revokes the old one.
    If a revoked refresh token is presented, all sessions for that user are immediately revoked
    (theft detection).

    ### Password reset (development)
...
  version: 0.8.0
  license:
    name: GNU General Public License v3.0
...
servers:
...
security:
  - CookieAuth: []
  - BearerAuth: []
...
paths:
...
  /api/folders/{folderID}/versions:
    get:
      summary: List all preserved versions of a file
      operationId: getFileVersions
      tags: [Versions]
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: folderID
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: path
          in: query
          required: true
          schema: { type: string }
          description: POSIX relative path of the file (e.g. "photos/img.jpg")
      responses:
        "200":
          description: List of versions, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileVersionsResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/folders/{folderID}/versions/{versionID}:
    get:
      summary: Download a specific file version
      operationId: getFileVersionDownload
      tags: [Versions]
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: folderID
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: versionID
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Raw file bytes for the requested version
          headers:
            X-Backup-Version:  { schema: { type: integer } }
            X-Checksum-SHA256: { schema: { type: string } }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/folders/{folderID}/backup/{relative_path}:
    parameters:
      - name: folderID
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: relative_path
        in: path
        required: true
        schema:
          type: string
        description: |
          POSIX-style path relative to the watched root (e.g. `photos/2024/img.jpg`).
          Each segment is percent-encoded; slashes are preserved. Must not contain `..`.
        example: photos/2024/img.jpg
    put:
      summary: Upload a file for backup
      description: |
        Streams the request body into object storage.

        **Deduplication:** if the `X-Checksum-SHA256` header matches the stored checksum
        for this path, the upload is skipped and `skipped: true` is returned.
      operationId: uploadFolderFile
      tags: [Backup]
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: X-Checksum-SHA256
          in: header
          required: true
          schema:
            type: string
          description: Hex-encoded SHA-256 of the file bytes.
        - name: X-File-Size
          in: header
          required: true
          schema:
            type: integer
          description: Exact byte count of the file.
        - name: X-Restored-From-Version-ID
          in: header
          required: false
          schema:
            type: integer
            format: int64
          description: >
            When this upload is a re-backup of a locally-restored file, set this
            to the ID of the version that was restored. The new version row will
            record that provenance in `restored_from_version_id`.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Upload complete (or skipped — see `skipped` field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadFileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: Storage or database failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: Download a backed-up file
      description: Streams the backed-up file bytes from object storage.
      operationId: downloadFolderFile
      tags: [Backup]
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: File bytes
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'attachment; filename="<name>"'
            Content-Length:
              schema:
                type: integer
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: DB record exists but the object is missing from storage (orphaned record)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a file's cloud backup
      description: |
        Permanently removes the backup record and **all versions** of the specified file from
        object storage and the database. The local file is not affected.
      operationId: deleteFileBackup
      tags: [Backup]
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        "204":
          description: Backup deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: Database or storage failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/history:
    get:
      summary: Get backup activity history
      description: |
        Returns file backup records sorted by `backed_up_at` DESC, enriched with the
        parent folder name and path. Supports filtering by folder and pagination.
      operationId: getBackupHistory
      tags: [History]
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: folder_id
          in: query
          required: false
          schema:
            type: integer
          description: Filter to a specific folder (omit for all folders)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 200
          description: Number of items to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Number of items to skip (for pagination)
      responses:
        "200":
          description: Paginated history items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        **Primary auth mechanism.** HttpOnly `access_token` cookie (SameSite=Lax, 1 minute TTL)
        set automatically by login/register/refresh. Insomnia captures it via the cookie jar.
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Fallback for clients that cannot use cookies (e.g. Electron from a file:// origin).
        Pass the `access_token` value from the login response body as `Authorization: Bearer <token>`.
  schemas:
    FileVersion:
      type: object
      required: [id, version, size, checksum_sha256, object_key, backed_up_at]
      properties:
        id:             { type: integer, format: int64 }
        version:        { type: integer }
        size:           { type: integer, format: int64 }
        checksum_sha256: { type: string }
        object_key:     { type: string }
        backed_up_at:   { type: string, format: date-time }
        restored_from_version_id:
          type: integer
          format: int64
          nullable: true
          description: >
            If this version was created by restoring a previous version, this
            field holds the ID of the source version. Null for versions that
            were created by a normal backup.
        delta_base_version_id:
          type: integer
          format: int64
          nullable: true
          description: >
            If non-null, the version's stored object is a binary bsdiff patch
            rather than a full copy. Reconstructing the file requires fetching
            the base version's bytes and applying the patch. Null for versions
            stored as full blobs (keyframes and first version of each file).
    FileVersionsResponse:
      type: object
      required: [versions]
      properties:
        versions:
          type: array
          items:
            $ref: "#/components/schemas/FileVersion"
    HistoryItem:
      type: object
      required: [id, user_id, watched_path_id, relative_path, size, checksum_sha256, object_key, backed_up_at, version, folder_name, folder_path]
      properties:
        id:          { type: integer, format: int64 }
        user_id:     { type: integer, format: int64 }
        watched_path_id: { type: integer, format: int64 }
        relative_path:   { type: string, example: "photos/2024/img.jpg" }
        size:        { type: integer, format: int64 }
        checksum_sha256: { type: string }
        object_key:  { type: string }
        backed_up_at: { type: string, format: date-time }
        version:     { type: integer }
        folder_name: { type: string }
        folder_path: { type: string }
    HistoryResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        total:  { type: integer }
        limit:  { type: integer }
        offset: { type: integer }
...
    UploadFileResponse:
      type: object
      required: [relative_path, size, checksum_sha256, backed_up_at, version, skipped]
      properties:
        relative_path:
          type: string
          example: photos/2024/img.jpg
        size:
          type: integer
          format: int64
        checksum_sha256:
          type: string
        backed_up_at:
          type: string
          format: date-time
        version:
          type: integer
          description: Increments on each content change; unchanged on a skipped upload.
          example: 3
        skipped:
          type: boolean
          description: true when the checksum matched and no upload occurred
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid credentials
  responses:
    BadRequest:
      description: Invalid request body or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"